Create lsst_sphinx_technote_theme package
I've moved the metadata-based configuration out of documenteer into the
theme's python package itself.

This is a good idea since it more strongly couples configuration to the
theme, while documenteer is a left as a general-purpose set of sphinx
utilities for enhancing LSST docs.

I expect that software docs will take the same route.