Create layout.html; just the boilerplate
Starting out the layout.html template with the minimum HTML5. Use
html5boilerplat and the sphinx_rtd_theme for inspiration.

My plan is to severely limit the amount of content that Sphinx adds to
the HTML to see if we an build better HTML and CSS with gulp. To improve
the body html, we're looking into implementing a custom html builder
with sphinx/docutils.