Add browersync support
Browsersync provides a server for viewing the site locally and on other
devices.

The sass task reloads browsersync; other tasks that provide content will
also need to trigger a reload.

A big limitation though is that Sphinx needs to actually be built for
the changes to show up in the demo site. The next step is to run the
Sphinx Makefile to trigger a build of the demo site.